MTSDK-416 Error Handling
- Revert to use Long for LogMessages.receiveMessageError function error.

+++ transport/src/androidUnitTest/kotlin/transport/shyrka/receive/ResponsesTests.kt
@@ -317,7 +317,7 @@ class ResponsesTests {
         decoded.run {
             assertThat(message).isEqualTo(ErrorTest.MESSAGE)
             assertThat(code).isEqualTo(PushTestValues.PUSH_CODE_DEPLOYMENT_NOT_FOUND)
-            assertThat(status).isEqualTo(ErrorTest.CODE_404)
+            assertThat(status).isEqualTo(ErrorTest.CODE_404.toInt())
             assertThat(contextId).isEqualTo(TestValues.DEFAULT_STRING)
             assertThat(details).isEmpty()
         }
